Fix backend and tests, etc.

+++ common/api/src/commonMain/kotlin/com/denchic45/stuiversity/api/course/work/submission/model/Grade.kt
@@ -25,7 +25,7 @@ data class Grade(
 
 @Serializable
 data class SubmissionGrade(
-    val value: Short,
+    val value: Int,
     @Serializable(UUIDSerializer::class)
     val courseId: UUID,
     @Serializable(UUIDSerializer::class)
